Break  into a helper where it can be unit tested in isolation
